requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
streamlit
youtube-transcript-api
python-dotenv